requirements.txt
polib>=1.1.1
opencc-python-reimplemented>=0.1.7
